+++ spec/support/mock_auth_helper.rb
